Build: Fixes for new-machine installation.

+++ install/install.sh
@@ -16,7 +16,7 @@ fi
 echo;while true; do
     read -p "Is SSH setup and working using 'ssh git@gitlab.com'. Continue? (y[es]/n[o]) " yn
     case $yn in
-        [Yy]* ) break;;
+        [Yy]* ) echo; break;;
         [Nn]* ) echo; echo "You're gonna have to be able to clone dotfiles for this to work ;-)"; exit;;
         * ) echo "Please answer yes or no.";
     esac
@@ -28,18 +28,26 @@ if [ ! -f /bin/zsh ] && [ ! -f /usr/bin/zsh ] && [ ! -f /usr/local/bin/zsh ]; th
     exit
 fi
 
-if [[ ! "$SHELL" =~ zsh$ ]]; then
+if [ "$1" != "noshellcheck" ] && [[ ! "$SHELL" =~ zsh$ ]]; then
     echo "Setting default shell to ZSH."
     chsh -s $(which zsh)
 
+    echo "If you are using Terminal.app, update preferences to launch /bin/zsh for 'Shell opens with'."
+    read -p "Hit ENTER when complete."
+
     echo "Relaunching in ZSH shell..."
-    zsh $0
+    zsh -c -i "${0} noshellcheck"
     exit
 fi
 
 #OhMyZsh
-if [ ! -d ~/.oh-my-zsh ]; then
-    sh -c "$(curl -fsSL https://raw.githubusercontent.com/robbyrussell/oh-my-zsh/master/tools/install.sh)"
+if [ ! -n "$OHMYZSH" ]; then
+    OHMYZSH=~/.oh-my-zsh
+fi
+
+if [ ! -d $OHMYZSH ]; then
+    umask g-w,o-w
+    git clone --depth=1 https://github.com/robbyrussell/oh-my-zsh.git "$OHMYZSH"
 fi
 
 #Code
@@ -53,7 +61,7 @@ fi
 
 [ ! -d ~/.dotfiles ] && ln -s ~/code/dotfiles ~/.dotfiles
 [ ! -f ~/.tmux.conf ] && ln -s ~/.dotfiles/tmux/tmux.conf.symlink ~/.tmux.conf
-[ ! -d ~/.vim ] && ln -s ~/.dotfiles/vim ~/.vim
+#[ ! -d ~/.vim ] && ln -s ~/.dotfiles/vim ~/.vim
 [ ! -d ~/.weechat ] && ln -s ~/.dotfiles/weechat ~/.weechat
 
 [ ! -d ~/.config ] && mkdir -p ~/.config
@@ -61,18 +69,14 @@ fi
 [ ! -d ~/.config/karabiner ] && ln -s ~/.dotfiles/karabiner ~/.config/karabiner
 [ ! -d ~/.config/nvim ] && ln -s ~/.dotfiles/nvim ~/.config/nvim
 
-if [ ! -f ~/.zshrc ]; then
-  ln -s ~/.dotfiles/zsh/zshrc.symlink ~/.zshrc
-  echo "Please reload ZSH shell with 'source ~./.zshrc;$0' to continue."
-  exit
-fi
+[ ! -f ~/.zshrc ] && ln -s ~/.dotfiles/zsh/zshrc.symlink ~/.zshrc
 
 #Karabiner-Elements
 if [ ! -d /Applications/Karabiner-Elements.app ]; then
     echo;while true; do
         read -p "Have you installed Karabiner-Elements? ([y]es/[n]o/[i]gnore) " yni
         case $yni in
-            [YyIi]* ) break;;
+            [YyIi]* ) echo; break;;
             [Nn]* ) echo; echo "Browse to 'https://pqrs.org/osx/karabiner' and install it."; exit;;
             * ) echo "Plase answer yes, no, or ignore."
         esac
@@ -84,18 +88,32 @@ if [ ! $(ls /Applications/Alfred*.app 2> /dev/null | wc -l) != "0" ]; then
     echo;while true; do
         read -p "Have you installed Alfred (and re-assigned shortcut to Cmd-Space)? ([y]es/[n]o/[i]gnore) " yni
         case $yni in
-            [YyIi]* ) break;;
+            [YyIi]* ) echo; break;;
             [Nn]* ) echo; echo "Browse to 'https://www.alfredapp.com' and install it."; exit;;
             * ) echo "Please answer yes, no, or ignore."s
         esac
     done
 fi
 
+#Colour Profiles (used primarily for tmux)
+function tic_(){
+    echo "TermInfo: ${1}"
+
+    if [[ ! $(toe) =~ (^|\ |
+)"${1}"\ * ]]; then
+        eval "tic -x ~/.dotfiles/terminfo/${1}.terminfo"
+    fi
+}
+
+tic_ tmux-256color
+tic_ xterm-256color
+
 #Brew
 if [ ! -x /usr/local/bin/brew ]; then
     /usr/bin/ruby -e "$(curl -fsSL https://raw.githubusercontent.com/Homebrew/install/master/install)"
-    munge_path /usr/local/bin
 fi
+
+munge_path /usr/local/bin
 
 echo;
 echo "Fetching Brew Taps...";BREW_TAPS="$(brew tap)"
@@ -110,6 +128,9 @@ function brew_install(){
 
     if [[ ! "$(brew ls --versions $FORMULA)" =~ ^($PKG^|$PKG\ $VERSION\.*|$FORMULA\ $VERSION\.*) ]]; then
         eval "brew install $@"
+        true
+    else
+        false
     fi
 }
     
@@ -140,64 +161,35 @@ brew_tap alecthomas/homebrew-tap
 brew_cask_install font-source-code-pro
 brew_cask_install font-source-code-pro-for-powerline
 
-#Languages
-brew_install python@2
-brew_install python@3
-brew_install ruby
-brew_install go
-brew_install gometalinter
-brew_install shellcheck
-brew_install yamllint
+#Misc
 brew_install tmux
+brew_install the_silver_searcher
+brew_install reattach-to-user-namespace
+brew_install fzf
+brew_install exa
+brew_install bat
+brew_install jq
+brew_install gnupg
+brew_install aspell
+brew_install weechat --with-aspell --with-curl --with-python@2 --with-perl --with-ruby --with-lua --with-guile
 
-#Rust
-[[ ! "$(rustc --version)" =~ ^rustc ]] && curl https://sh.rustup.rs -sSf | sh
-
+#Package helpers
 function cargo_install(){
-    echo "Rust: ${1} (cargo)"
+    echo "Rust: $1 (cargo)"
 
     if [ ! -x ~/.cargo/bin/${1} ]; then
-        eval "cargo install ${1} ${2}"
-    fi
-
-    munge_path ~/.cargo/bin
-}
-
-cargo_install cargo-tree +stable
-cargo_install cargo-outdated +stable
-cargo_install racer +nightly
-
-#Go
-#[ ! -f "$GOPATH/bin/gometalinter" ] && go get -u github.com/nsf/gocode
-
-#Alacritty
-if [ ! -d /Applications/Alacritty.app ]; then
-    git clone https://github.com/jwilm/alacritty.git ~/code/alacritty
-    make app
-    cp -r target/release/osx/Alacritty.app /Applications
-fi
-
-#Colour Profiles
-function tic_(){
-    echo "TermInfo: ${1}"
-
-    if [[ ! $(toe) =~ (^|\ |
-)"${1}"\ * ]]; then
-        eval "tic -x ~/.dotfiles/terminfo/${1}.terminfo"
+        eval "cargo ${2} install ${1}"
     fi
 }
 
-tic_ tmux-256color
-tic_ xterm-256color
 
-#Neovim
 function pip_install()
 {
-    echo "pip${1}: ${2}${3}${4}${5} (package)"
+    echo "pip${1}: ${2} ${3} ${4} ${5} (package)"
 
     if [[ ! $(eval "pip${1} freeze") =~ (^|\ |
 )"${2}"==* ]]; then
-        eval "pip${1} ${2}${3}${4}${5}"
+        eval "pip${1} install ${2} ${3} ${4} ${5}"
     fi
 }
 
@@ -211,18 +203,62 @@ function gem_install()
     fi
 }
 
+#Languages
+brew_install python@2
+brew_install python@3
+brew_install go
+brew_install gometalinter
+brew_install shellcheck
+brew_install yamllint
+
+#Rust
+[ ! -x ~/.cargo/bin/rustc ] && curl https://sh.rustup.rs -sSf | sh
+
+munge_path ~/.cargo/bin
+
+[[ ! "$(rustup show)" =~ nightly\- ]] && rustup install nightly
+
+cargo_install cargo-tree
+cargo_install cargo-outdated
+cargo_install racer +nightly
+
+#Go
+#[ ! -f "$GOPATH/bin/gometalinter" ] && go get -u github.com/nsf/gocode
+
+#Ruby
+if brew_install rbenv; then
+    rbenv init
+    #eval("rbenv init -")
+
+    RUBY_LATEST=$(rbenv install -l | awk -F '.' '
+        /^[[:space:]]*[0-9]+\.[0-9]+\.[0-9]+[[:space:]]*$/ {
+            if ( ($1 * 100 + $2) * 100 + $3 > Max ) { 
+                Max = ($1 * 100 + $2) * 100 + $3
+                Version=$0
+            }
+        }
+        END { print Version }')
+
+    rbenv install $RUBY_LATEST
+    rbenv global $RUBY_LATEST
+fi
+
+#Alacritty
+if [ ! -d /Applications/Alacritty.app ]; then
+    if [ ! -d ~/code/alacritty ]; then
+        git clone https://github.com/jwilm/alacritty.git ~/code/alacritty
+    fi
+
+    pushd ~/code/alacritty
+    make app
+    cp -r target/release/osx/Alacritty.app /Applications
+    rm -rf target/release/osx/Alacritty.app
+    rm -rf target/assets/osx/Alacritty.app
+    popd
+fi
+
+#Neovim
 brew_install neovim
 pip_install 2 neovim
 pip_install 3 neovim --upgrade
 gem_install neovim
-
-#Misc
-brew_install the_silver_searcher
-brew_install reattach-to-user-namespace
-brew_install fzf
-brew_install exa
-brew_install bat
-brew_install jq
-brew_install gnupg
-brew_install aspell
-brew_install weechat --with-aspell --with-curl --with-python@2 --with-perl --with-ruby --with-lua --with-guile
